refactor(api): move identity access management script
